build: Update spring boot to 2.4.1
In order to update spring boot to the
current version 2.4.1 also various other
dependencies needed to be updated.
Including oracle database dependencies
and changing flyway to the integrated
version of spring boot.

Furthermore the repositories method 
changed in spring that's why this needed
adaption. Additionally the tests needed
to be adapted and deprecated method
removed.