Update to Java 11
To upgrade to Java 11 we had to update lombok and Mockito since the
application wouldn't build/the tests wouldn't run anymore. Mockito was
only updated to 2.25 since upgrading to 3.x.x would require changes
which are out of the scope at the moment.
The locale used in the RecordJoiner also had to be changed to
Locale.GERMANY because with Java 9 the default locale provider was
changed to CLDR. CLDR doesn't provide the start of the week for language
specific locales, so it would fallback to the default which is Sunday.
Our tests expected monday as the start of the week so those failed.
This fix is probably not optimal but good enough for now since we need
to work on internationalization in the future anyway.